requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
Flask-RESTX==1.3.0
WTForms==3.1.1
pytest==7.4.3
pytest-flask==1.3.0
ipaddress==1.0.23
coverage==7.3.2
gunicorn==21.2.0
python-dotenv==1.0.0